Make sure people see the printer dialog
Windows tends to open the print dialog *below* the current AM window. Depending
on the size of the window, the user does not even see the dialog. To make
matters worse, as we have disabled the UI of the PDF viewer, there is not even
an entry in the task list for the viewer, making it impossible for the user to
see what's going on.

So we need to bring the window to the foreground. The most obvious candidate in
the Windows API for this is SetForegroundWindow(), but Windows restricts when
this function may be called as it steals focus. The only other option is
forcing the window to show up on top of all other windows, which this code does
(this operation does *NOT* steal focus, it just makes the window very visible
and impossible to hide).

As we block on the call to the PDF viewer, we have to do the window elevation
in the background. Luckily, goroutines make this really easy. We just keep
looping over all windows with a short pause inbetween until we have found our
window.