Scale output of simple PCL jobs
I don't know how, but the PCL generated by AM for simple jobs doesn't fit onto
an DIN A4 page - it's too wide. Maybe it works in PCL mode because the margins
are ridiculously small, but GhostPCL swallows a few columns.

This patch adds a workaround: It first creates an intermediate PDF with a wider
paper format and then feeds that PDF through ghostscript again, shrinking it to
fit onto DIN A4.

I *think* that we only have to do this for jobs which do not contain PJL
commands, so that's what I use to switch the additonal step on and off,
because we definitely *don't* want to scale invoices etc.